refactoring commands
